+++ chain/x/pole/types/keys.go
@@ -20,4 +20,5 @@ var (
 	ReplicaReceiptsKeyPrefix   = collections.NewPrefix(9)
 	NodesKeyPrefix             = collections.NewPrefix(10)
 	VerificationRecordsKeyPrefix = collections.NewPrefix(11)
+	AnnualEmissionKeyPrefix    = collections.NewPrefix(12)
 )